refactor(FeatureCtx): move class FeatureContext to Style and rename
