+++ src/Converter/Feature2Texture.js
@@ -2,9 +2,9 @@ import * as THREE from 'three';
 import { FEATURE_TYPES } from 'Core/Feature';
 import Extent from 'Core/Geographic/Extent';
 import Coordinates from 'Core/Geographic/Coordinates';
-import { FeatureContext } from 'Converter/Feature2Mesh';
+import { StyleContext } from 'Core/Style';
 
-const context = new FeatureContext();
+const context = new StyleContext();
 
 const svg = document.createElementNS('http://www.w3.org/2000/svg', 'svg');
 const matrix = svg.createSVGMatrix();
